mydj/requirements.txt
django-filter
markdown
djangorestframework
django-cors-headers
pygments
djangorestframework-jwt